requirementsWin.txt
annotated-types==0.6.0
argcomplete==1.10.3
asgiref==3.7.2
attrs==23.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.8.2
blis==0.7.11
catalogue==2.0.10
certifi==2023.11.17
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
# cmake==3.28.3
colorama==0.4.6
compressed_rtf==1.0.6
confection==0.1.4
cryptography==41.0.7
cymem==2.0.8
defusedxml==0.8.0rc2
Django==3.2.5
django-cors-headers==4.3.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
# dlib==19.24.2
docx2txt==0.8
ebcdic==1.1.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
extract-msg==0.28.7
# face-recognition==1.3.0
# face-recognition-models==0.3.0
idna==3.6
IMAPClient==2.1.0
importlib-resources==6.1.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
langcodes==3.3.0
lxml==5.1.0
MarkupSafe==2.1.4
murmurhash==1.0.10
nltk==3.8.1
numpy==1.24.4
oauthlib==3.2.2
olefile==0.47
packaging==23.2
pandas==2.0.3
pdfminer.six==20191110
pillow==10.2.0
pkgutil_resolve_name==1.3.10
preshed==3.0.9
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
pyresparser==1.0.6
pyrsistent==0.20.0
python-dateutil==2.8.2
python-pptx==0.6.23
python3-openid==3.2.0
pytz==2023.3.post1
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.17.1
scikit-learn==1.3.2
scipy==1.10.1
six==1.12.0
smart-open==6.4.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.8.1
sqlparse==0.4.4
srsly==2.4.8
textdistance==4.6.1
textract==1.6.5
thinc==8.2.2
threadpoolctl==3.3.0
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
wasabi==1.1.2
weasel==0.3.4
xlrd==1.2.0
XlsxWriter==3.2.0
zipp==3.17.0
mysqlclient
gunicorn
#celery
# redis
# facenet_pytorch
# torch
# django-celery-results
#Django - 4.2.11
#django-redis
#boto3
#django-storages
#python-dotenv